@@ -17,8 +17,9 @@
 	},
 	"devDependencies": {
 		"@types/node": "^15.0.3",
-		"@types/react": "^17.0.5",
-		"@types/react-dom": "^17.0.5",
+		"@types/react": "18.0.9",
+		"@types/react-dom": "18.0.4",
+    	"typescript": "4.4.2",
 		"@types/react-router": "^5.1.18",
 		"@types/react-router-dom": "^5.3.3",
 		"css-loader": "^6.7.1",
@@ -27,7 +28,6 @@
 		"react-app-alias": "^2.1.0",
 		"serve": "11.3.2",
 		"ts-loader": "^9.1.2",
-		"typescript": "^4.2.4",
 		"webpack": "^5.67.0",
 		"webpack-cli": "^4.9.2",
 		"webpack-dev-server": "^4.7.3"
