@@ -4,14 +4,14 @@
 	"private": true,
 	"devDependencies": {
 		"@types/node": "^15.0.3",
-		"@types/react": "^17.0.5",
-		"@types/react-dom": "^17.0.5",
+		"@types/react": "18.0.9",
+		"@types/react-dom": "18.0.4",
 		"css-loader": "^6.7.1",
 		"html-webpack-plugin": "5.5.0",
 		"serve": "13.0.2",
 		"style-loader": "^3.3.1",
 		"ts-loader": "^9.1.2",
-		"typescript": "^4.2.4",
+		"typescript": "4.4.2",
 		"webpack": "^5.67.0",
 		"webpack-cli": "^4.9.2",
 		"webpack-dev-server": "^4.7.3"
