@@ -1,4 +1,7 @@
-import App from "./_app";
-import ReactDOM from "react-dom";
+import App from './_app';
+import * as ReactDOMClient from "react-dom/client";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const rootElement = document.getElementById("root");
+const root = ReactDOMClient.createRoot(rootElement!);
+
+root.render(<App />);
