@@ -1,9 +1,11 @@
-import React, { Suspense, useContext, useEffect } from "react";
+import React, { Suspense, useState } from "react";
 import { NavLink } from "react-router-dom";
 import PageLoader from "../components/page-loader";
+import LeftNavigation from "@shared/components/left-navigation";
+import Sidemenu from "./sidemenu";
+import Footer from "./footer";
 
-import '../assets/styles/bootstrap.css';
-import '../assets/styles/main.css';
+import '../assets/styles/main.scss';
 
 const Header = () => (
   <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
@@ -16,14 +18,29 @@ interface LayoutProps {
   children: React.ReactNode
 };
 
+const mql = window.matchMedia(`(min-width: 800px)`);
+
 const Layout = ({ children }: LayoutProps) => {
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+  const [searchOpen, setSearchOpen] = useState(false);
+  const [sidebarDocked, setSidebarDocked] = useState(mql.matches);
+
+  mql.addListener(() => {
+    setSidebarDocked(mql.matches);
+  });
+
   return (
     <Suspense fallback={<PageLoader />}>
       <Header />
-      <div className="container mt-4">{children}</div>
-      <div className="text-center text-muted">
-        cfryerdev - 2022
-      </div>
+      {/* <LeftNavigation
+        sidebar={<Sidemenu toggleMenu={() => setSidebarOpen(false)} />}
+        open={sidebarOpen}
+        onSetOpen={setSidebarOpen}
+        docked={sidebarDocked}
+      > */}
+        <div className="container mt-4">{children}</div>
+      {/* </LeftNavigation> */}
+      <Footer />
     </Suspense>
   );
 };
